--- src/firefox/backgroundScripts/background.js
@@ -1,6 +1,0 @@
-import { testShared } from "../../shared/testshared.js";
-
-browser.runtime.onInstalled.addListener(() => {
-  console.log("onInstalled...");
-  testShared();
-});
